Add previously non-git controlled files
We used to symlink the git controlled things into the non-git controlled
h-lame.com folder, but with all the changes to the .htaccess file we can
get everything out that's not essential and leave only these handful of
files left.

We could probably clean out some of these, but we'll deal with that later.