Add "red fox"
I wrote this weird tone poem way back in 2000 / 2001 when I first moved
to London and saw a dead fox squashed by a bus as I walked to Kilburn tube
station.  I don't think I'd seen a live fox in my life before so it was
a pretty horrid thing to see.

Speaking of horrid things to see, the HTML in this is pretty gross, but
it's impressive how much this still looks like it was supposed to.