Found a xmas "card" I made
Looks like this was done with some kind of flash timeline to dhtml
converter as the animation path on the card has functions that are
copyright macromedia and are prefixed with `MM_`.  I don't remember how
I generated this.

I only had to tweak the HTML and JS a tiny bit for it to keep working:

* Replacing `document.all[]` with `document.getElementById` (even though
  it still worked, it was in a ternary that relied on `document.all`
  returning true which it no longer does).
* Removing a pointless form element so forcing you to press "OK" - I tried
  to get an onSubmit handler to work, but it somehow lost the `opener.who`
  meaning your name wasn't passed to the card from the form.