Introduce slides concept for talks pages
All slides are now individual markdown documents with frontmatter for the
image urls and markdown for the slide text.  These are all ignored so not
rendered as individual pages, but used to programmatically build the table
of contents at the top of the page and then inline rendered via the
`_slide.erb` partial in the page.  The only content in the index file is
the frontmatter and any text to go into the "intro" paragraph at the top
of the page before we get to the talk itself.