Introduce `html_to_slides` script to convert talks html to slides/*.md
Luckily we have a fairly well structured format and can pull it apart with
nokogiri to turn it from one html file to several slides/*.md files.

This first version of the script extracts the details of the slide image,
and adds a new `image_title` attribute while we're at it, and then pulls
the rest of the transacript as raw HTML into the slide.  Future commits
will do their best to convert the transcript HTML into markdown.