Stop outputting so many newlines
All our files have trailing newlines (this is the way) so we get lots of
extra newlines in the html once we render files inside files.  Obvs, the
indentation is _aaaallll_ over the place so it shouldn't matter too much,
but we can avoid the extra newlines with the `-%>` erb closing tag, so we
might as well.