Extract "terminal output" sections from HTML slides
Some of the "slides" in the data talk are, in fact, terminal output that's
been marked up in HTML and not images at all.  Support extracting that
from the HTML version as a chunk of markdown in the frontmatter and then
rendering it instead of the image slide in the slide template.

The terminal output isn't really usefully markdownable, nor do I think is
it much simpler to write an erb helper for it, so it's just a chunk of
HTML that we render directly.