Tweak results of using typebase on talks css
The model for typebase seems to be using margin-top to separate elements
and having everything with no margin-bottom.  That doesn't' _quite_ fit
with how we were doing things before.  We introduce some extra helper
classes to add to the existing .hug class that typebase provides so we can
force or reject margin-top and margin-bottom if we need to.