Make slide titles more semantic
As the slides are in section tags they should probably have heading tags
to identify them.  We can apply the same formatting to some h2s as we
apply to the title attribute to move them into the border.  This should
mean the markup is more semantic, and still renders the same.

We also put h2s on any other sections, like the intro, or videos and then
hide those if they're not wanted.