Upload release asset to tag
Following the lead of how we do things on lrug/lrug.org we build the site
then zip it up and upload it as an asset to a single "deploy" release on
github.  This gives us a known url to download it from with the deploy
webhook.  We'll deal with pinging the deploy webhook later.