Add anchor links to headings
Stealing some more code from the LRUG site, we augment the kramdown
renderer for headings to add a <span> with an <a> pointing to the id of
the header at the start of the element.  At the start because I want the
links in the marginalia on the left.

Unlike LRUG we allow for turning the anchors off because, for example, on
the talks pages the heading _already_ have links to the bits of the page
the headers are generating ids for.  We can do this with
`heading_anchors: false` in the frontmatter.