Adding `scripts/publish.rb` to automatically add `published_at`
For now it only looks at `notes/` as that's all that's gonna go in the
feed to begin with.  Loops through the resources and adds `published_at`
to the frontmatter if it's not present.  Will massage the frontmatter on
write which is a little annoying, but ruby doesn't give us many options
for programmatically saying how we want yaml output to be formatted.