Update build+deploy action to "publish" new notes
We run our `scripts/publish` file and then if that caused any changes we
commit them and push.  By design github actions don't get caught in a
infinite loop if the actions themselves cause commits / pushes that might
otherwise trigger the action to start, so this is safe.

All this so I can write notes without having to decide when they'll be
published, and on push they'll get a timestamp and this can then be used
to populate a feed.