Added :active and :target styling
It seems :active might be marginally nicer for touch devices, and :target
will help with identifying what you've been deeplinked into.  Would be
nice to highlight the whole section, but I've not wrapped my headings and
their subsequent content in a container, so it makes it harder.  I did try
a:

    *:target {
      // target heading style effect

      ~ :not([id]) {
        // target content style
      }
      ~ [id] {
        ~ :not([id]) {
          // content for subsequent headings style reset
        }
      }
    }

effect, but it seemed a bit much, and I couldn't quite work out what
styling I even wanted.