Remove a redundant markdown render from footnotes
The whole doc is markdown (probably) so we'll get the render from the
whole doc and don't need to markdown render each footnote.  This has the
added advantage of moving any abbreviations or link definitions out of the
individual footnotes themselves and into the footnote section as a whole.
It'd be nice to work out how to have them live next to the definitions for
the whole doc, but this'll do for now.