Adding an inline image list
Lots of fancy responsive stuff going on here, but this should mean on
mobile we show 600px wide images full width, and on desktop we show 190px
wide images 3-up.  Which is weird obvs because mobile gets the _bigger_
image and takes the download hit, but ... eh?  We've also set 2x and 1x
images for both so retina screens should do the nicest thing.  Happily,
the 1x 600px image is a good substitute for 3x on desktop.  I've not
bothered with a 3x for mobile.