Add and configure middleman-syntax for syntax highlighting of code
Watch out, I'm gonna write some code snippets in some future blog post and I'm gonna want them to look nice.  I've "hardcoded" the colour scheme because the advice to use `erb` to generate one on the fly didn't play nice with my use of `scss` and `@import` in the main css.  I didn't want to add a separate css file just for it, so I generated it on the fly as a standalone css file using erb and then copy and pasted it into a scss file that I _could_ `@import`.  Faff, but we're done.