Re-order "slides" for the talks index
I had 1->6 but realised that to add a new one that I wanted at the top of the I'd need to move the existing 1->6 to 2->7 and then add a new 1.  This was obvs gonna be a massive pain for each talk I added so I added the ability to tell the index to order the slides in reverse - I did have to reverse the contents of the existing 1->6 to be the contents of 6->1, but that's at least a one time hit.