Prepare for merging the "dynamic" branch
This "dynamic" branch is where all the real work lives, so let's merge it to the default branch, but while we're at it, let's not have the default branch be called "master" anymore - "main" is where it's at.

In the repo this means updating the deploy script (which needs to be manually deployed) and the github actions config to stop talking about the "dynamic" branch.  Outside the repo this means doing the "default branch rename" in github and then merging the [PR for the dynamic branch][dynamic-pr]

[dynamic-pr]: https://github.com/h-lame/h-lame.com/pull/1